+++ src/main/java/org/example/tablenow/domain/event/message/producer/EventOpenProducer.java
@@ -1,9 +1,9 @@
-package org.example.tablenow.global.rabbitmq.event.producer;
+package org.example.tablenow.domain.event.message.producer;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.tablenow.global.rabbitmq.constant.RabbitConstant;
-import org.example.tablenow.global.rabbitmq.event.dto.EventOpenMessage;
+import org.example.tablenow.domain.event.message.dto.EventOpenMessage;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Component;
