+++ src/main/java/org/example/tablenow/domain/event/service/EventService.java
@@ -14,8 +14,8 @@ import org.example.tablenow.domain.store.entity.Store;
 import org.example.tablenow.domain.store.service.StoreService;
 import org.example.tablenow.global.exception.ErrorCode;
 import org.example.tablenow.global.exception.HandledException;
-import org.example.tablenow.global.rabbitmq.event.dto.EventOpenMessage;
-import org.example.tablenow.global.rabbitmq.event.producer.EventOpenProducer;
+import org.example.tablenow.domain.event.message.dto.EventOpenMessage;
+import org.example.tablenow.domain.event.message.producer.EventOpenProducer;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
